refactor(data): move index builders to store module
